+++ app/src/test/resources/db/changelog/changes/V1.0__Init.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS document_summary (
     id varchar(255) NOT NULL,
-    number integer,
+    number BIGINT,
     author varchar(255),
     title varchar(255),
     create_date timestamp without time zone,
